feat: finalizado rota GET

+++ src/schemas/movie.schemas.ts
@@ -4,8 +4,8 @@ const movieSchema = z.object({
     id: z.number().positive(),
     name: z.string().max(50),
     description: z.string().nullish(),
-    duration: z.number().positive(),
-    price: z.number().positive()
+    duration: z.number().int().positive(),
+    price: z.number().int().positive()
 });
 
 const movieCreateSchema = movieSchema.omit({ id: true });
